.github/tests/acceptance/test_arduino_challenge/requirements.txt
pytest
rpi-lgpio
smbus2
pyserial
numpy
matplotlib
